+++ CHANGELOG.rst
@@ -29,6 +29,7 @@ Added
   to specify the host on which administrative UI and API will be opened.
 - Refactor two-phase commit logics: don't use hardcoded timeout value for the
   ``prepare`` stage, move ``upload`` to a separate stage.
+- Allow operating cluster from an unconfigured instance WebUI.
 
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 Fixed